Normalization: remove ssh errors if ssh is not being used (#6504)
* add check for ssh config file and change exits to returns

* bump normalization version

* bump version

* version de-bump

+++ airbyte-workers/src/main/resources/sshtunneling.sh
@@ -2,14 +2,19 @@
 # Requires one argument,
   # path to ssh config file ($1)
 function openssh() {
+  # check if the ssh config file exists, it won't if ssh is off
+  if [ ! -f $1 ]; then
+    echo "detected no config file for ssh, assuming ssh is off."
+    return 0
+  fi
   # check if jq is missing, and if so try to install it..
   # this is janky but for custom dbt transform we can't be sure jq is installed as using user docker image
   if ! command -v jq &> /dev/null ; then
-    echo "CRITICAL: jq not installed... attempting to install on the fly but will fail if unable."
+    echo "CRITICAL: jq not installed... attempting to install on the fly."
     { apt-get update && apt-get install -y jq; } ||
     apk --update add jq ||
     { yum install epel-release -y && yum install jq -y; } ||
-    { dnf install epel-release -y && dnf install jq -y; } || exit 1
+    { dnf install epel-release -y && dnf install jq -y; } || return 0
   fi
   # tunnel_db_host and tunnel_db_port currently rely on the destination's spec using "host" and "port" as keys for these values
   # if adding ssh support for a new destination where this is not the case, extra logic will be needed to capture these dynamically
@@ -25,7 +30,7 @@ function openssh() {
     echo "Detected tunnel method SSH_KEY_AUTH for normalization"
     # create a temporary file to hold ssh key and trap to delete on EXIT
     trap 'rm -f "$tmpkeyfile"' EXIT
-    tmpkeyfile=$(mktemp /tmp/xyzfile.XXXXXXXXXXX) || exit 1
+    tmpkeyfile=$(mktemp /tmp/xyzfile.XXXXXXXXXXX) || return 1
     echo "$(cat $1 | jq -r '.tunnel_map.ssh_key')" > $tmpkeyfile
     # -f=background  -N=no remote command  -M=master mode  StrictHostKeyChecking=no auto-adds host
     echo "Running: ssh -f -N -M -o StrictHostKeyChecking=no -S {control socket} -i {key file} -l ${tunnel_username} -L ${tunnel_local_port}:${tunnel_db_host}:${tunnel_db_port} ${tunnel_host}"
@@ -40,7 +45,7 @@ function openssh() {
       { apt-get update && apt-get install -y sshpass; } ||
       { apk add --update openssh && apk --update add sshpass; } ||
       { yum install epel-release -y && yum install sshpass -y; } ||
-      { dnf install epel-release -y && dnf install sshpass -y; } || exit 1
+      { dnf install epel-release -y && dnf install sshpass -y; } || return 1
     fi
     # put ssh password in env var for use in sshpass. Better than directly passing with -p
     export SSHPASS=$(cat $1 | jq -r '.tunnel_map.tunnel_user_password')
